Power optimization: Scan screen only when user touches it
The current implementation scans the screen every 750 milliseconds. This
is pretty expensive in terms of power consumption.

Since the screen mode (map, gym, pokemon, etc) doesn't change unless the
user actually touches the screen, we just need to scan the screen around
the time the screen is touched.

Also, since the user touches the screen a lot during a gym battle, add
some throttling to scan the screen only after the last touch if multiple
touches happen in quick succession (within 750 milliseconds).

This should significantly improve the power.

Since there are some animations which take time when the user opens a
pokemon or switches pokemon, scan a couple of times before stopping.